Switch some EDiag()s to DDiag()s.

+++ runtime/gc/foster_gc.cpp
@@ -242,9 +242,10 @@ class copying_gc {
           if (arr) {
             cell_size = array_size_for(arr->num_elts(), map->cell_size);
             if (ENABLE_GCLOG) {
-              fprintf(gclog, "Collecting array of total size %lld (rounded up from %ld + %lld = %lld), cell size %lld, len %lld...\n",
+              fprintf(gclog, "Collecting array of total size %lld (rounded up from %lld + %lld = %lld), cell size %lld, len %lld...\n",
                                   cell_size,
-                                  sizeof(heap_array),  arr->num_elts() * map->cell_size,
+                                  int64_t(sizeof(heap_array)),
+                                                       arr->num_elts() * map->cell_size,
                                   sizeof(heap_array) + arr->num_elts() * map->cell_size,
                                   map->cell_size,
                                   arr->num_elts());
